requirements.txt
construct==2.5.3
Django==1.11.6
django-haystack==2.6.1
django-markdownx==2.0.21
future==0.16.0
jieba==0.39
Markdown==2.6.9
olefile==0.44
pefile==2017.9.3
Pillow==4.3.0
python-ptrace==0.9.3
pytz==2017.2
six==1.11.0
Whoosh==2.7.4
